Slightly refactor first two policy tests
Improve naming, fix some comments
