+++ tck/app-custom-policy/src/main/webapp/WEB-INF/beans.xml
