+++ tck/app-custom-policy2/src/main/webapp/WEB-INF/web.xml
