--- tck/app-policy2/pom.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!--
-
-    Copyright (c) 2024 Contributors to the Eclipse Foundation.
-    Copyright (c) 2015, 2020 Oracle and/or its affiliates. All rights reserved.
-
-    This program and the accompanying materials are made available under the
-    terms of the Eclipse Public License v. 2.0, which is available at
-    http://www.eclipse.org/legal/epl-2.0.
-
-    This Source Code may also be made available under the following Secondary
-    Licenses when the conditions for such availability set forth in the
-    Eclipse Public License v. 2.0 are satisfied: GNU General Public License,
-    version 2 with the GNU Classpath Exception, which is available at
-    https://www.gnu.org/software/classpath/license.html.
-
-    SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
-
--->
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-        <groupId>org.eclipse.ee4j.authorization.tck</groupId>
-        <artifactId>jakarta-authorization-tck</artifactId>
-        <version>4.0.0-SNAPSHOT</version>
-    </parent>
-
-	<artifactId>app-mem-policy2</artifactId>
-	<packaging>war</packaging>
-	
-	<description>
-        This tests executing a custom permission check before authentication takes place.
-    </description>
-
-	<properties>
-        <failOnMissingWebXml>false</failOnMissingWebXml>
-    </properties>
-
-	<dependencies>
-        <dependency>
-            <groupId>org.eclipse.ee4j.authorization.tck</groupId>
-            <artifactId>common</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-
-	<build>
-        <finalName>app-mem-policy2</finalName>
-	</build>
-</project>
